requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
configparser==7.0.1
dnspython==2.6.1
dotenv==0.9.9
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipaddress==1.0.23
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3

openai~=1.78.1